form creation fixes
