+++ forms-gen-nest/apps/forms-db/src/dto/generation_from_complete.dto.ts
@@ -1,13 +1,31 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsInt, IsString } from 'class-validator';
+import { Type } from 'class-transformer';
+import { IsArray, IsOptional, IsString, ValidateNested } from 'class-validator';
 
-export class GenerationUpdateDto {
+export class GenerationCompleteDto {
     @ApiProperty()
     @IsString()
     id: string;
 
-    @ApiProperty()
-    @IsInt()
-    questions_done: number;
+    @ApiProperty( {type: () => [GenerationCompleteQuestionDto] } )
+    @IsArray()
+    @Type( () => GenerationCompleteQuestionDto )
+    @ValidateNested( { each: true } )
+    questions: Array<GenerationCompleteQuestionDto>;
 }
 
+export class GenerationCompleteQuestionDto {
+    @ApiProperty()
+    @IsString()
+    question: string;
+
+    @ApiProperty( {type: () => [String] } )
+    @Type( () => String )
+    @IsArray()
+    answers: Array<string>;
+
+    @ApiProperty()
+    @IsString()
+    @IsOptional()
+    correct_answer?: string;
+}
