+++ forms-gen-nest/apps/forms-db/src/dto/generation_from_update.dto.ts
@@ -1,31 +1,12 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { Type } from 'class-transformer';
-import { IsArray, IsOptional, IsString, ValidateNested } from 'class-validator';
+import { IsInt, IsString } from 'class-validator';
 
-export class GenerationCompleteDto {
+export class GenerationUpdateDto {
     @ApiProperty()
     @IsString()
     id: string;
 
-    @ApiProperty( {type: () => [GenerationCompleteQuestionDto] } )
-    @IsArray()
-    @Type( () => GenerationCompleteQuestionDto )
-    @ValidateNested( { each: true } )
-    questions: Array<GenerationCompleteQuestionDto>;
-}
-
-export class GenerationCompleteQuestionDto {
     @ApiProperty()
-    @IsString()
-    question: string;
-
-    @ApiProperty( {type: () => [String] } )
-    @Type( () => String )
-    @IsArray()
-    answers: Array<string>;
-
-    @ApiProperty()
-    @IsString()
-    @IsOptional()
-    correct_answer?: string;
+    @IsInt()
+    questions_done: number;
 }
